Split dwd source adapter into multiple classes.
