+++ server/src/main/java/org/jvalue/ods/processor/adapter/domain/weather/dwd/DwdWeatherServiceSourceAdapter.java
@@ -1,7 +1,6 @@
-package org.jvalue.ods.processor.adapter.domain.weather;
+package org.jvalue.ods.processor.adapter.domain.weather.dwd;
 
 import com.codahale.metrics.MetricRegistry;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.inject.assistedinject.Assisted;
 import org.jvalue.ods.api.sources.DataSource;
@@ -9,7 +8,7 @@ import org.jvalue.ods.processor.adapter.JsonSourceIterator;
 import org.jvalue.ods.processor.adapter.SourceAdapter;
 import org.jvalue.ods.processor.adapter.SourceAdapterException;
 import org.jvalue.ods.processor.adapter.SourceAdapterFactory;
-import org.jvalue.ods.processor.adapter.domain.weather.models.*;
+import org.jvalue.ods.processor.adapter.domain.weather.models.Location;
 import org.jvalue.ods.utils.JsonMapper;
 
 import javax.annotation.Nonnull;
@@ -18,9 +17,6 @@ import javax.ws.rs.core.UriBuilder;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URL;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 
@@ -37,8 +33,6 @@ public class DwdWeatherServiceSourceAdapter implements SourceAdapter {
 	private final NodeParsingStrategy responseParser;
 
 	/**
-	 * Only uses the first location. No need for more than one. TODO make the user only input one location.
-	 *
 	 * @throws IllegalArgumentException if no Location is specified or time is neither current nor forecast.
 	 */
 	@Inject
@@ -82,77 +76,6 @@ public class DwdWeatherServiceSourceAdapter implements SourceAdapter {
 	}
 
 	/**
-	 * Interface for current and forecast response parsing strategies.
-	 */
-	private interface NodeParsingStrategy {
-		Iterator<ObjectNode> parseServiceResponse(Iterator<ObjectNode> nodeIterator);
-	}
-
-	/**
-	 * NodeParsingStrategy for current time.
-	 */
-	private class CurrentResponseParser implements NodeParsingStrategy {
-		public Iterator<ObjectNode> parseServiceResponse(Iterator<ObjectNode> nodeIterator) {
-			if (!nodeIterator.hasNext()) return Collections.emptyIterator();
-
-			ObjectNode node = nodeIterator.next();
-
-			String stationId = "unknown";
-			Instant timestamp = Instant.parse(node.get("time").asText());
-
-			JsonNode weatherNode = node.get("weather");
-
-			JsonNode temperatureNode = findDataPointNodeByName(weatherNode, "temperature200");
-			Temperature temperature = null;
-			if (temperatureNode != null) {
-				double temperatureValue = temperatureNode.get("value").asDouble();
-				temperature = new Temperature(TemperatureType.CELSIUS.fromKelvin(temperatureValue), TemperatureType.CELSIUS);
-			}
-
-			JsonNode airPressureNode = findDataPointNodeByName(weatherNode, "air_pressure");
-			Pressure pressure = null;
-			if (airPressureNode != null) {
-				double pressureValue = airPressureNode.get("value").asDouble();
-				pressure = new Pressure((int) pressureValue, PressureType.H_PA);
-			}
-
-			//TODO this is currently not delivered in current, but could be if the service differentiates between current and forecast standard parameters.
-			JsonNode humidityNode = findDataPointNodeByName(weatherNode, "humidity");
-			int humidityInPercent = -1;
-			if (humidityNode != null) {
-				double humidityValue = humidityNode.get("value").asDouble();
-				humidityInPercent = (int) Math.round(humidityValue);
-			}
-
-			Weather weather = new Weather(//TODO extend the weather model to support all standard parameters
-				stationId,
-				temperature,
-				pressure,
-				humidityInPercent,
-				timestamp,
-				location);
-
-			ObjectNode resultNode = JsonMapper.valueToTree(weather);
-			ArrayList<ObjectNode> resultList = new ArrayList<>();
-			resultList.add(resultNode);
-			return resultList.iterator();
-		}
-
-	}
-
-	/**
-	 * @return the JsonNode where the "name" field is set to name.
-	 */
-	private JsonNode findDataPointNodeByName(JsonNode node, String name) {
-		for (JsonNode subNode : node) {
-			if (name.equals(subNode.get("name").asText())) {
-				return subNode;
-			}
-		}
-		return null;
-	}
-
-	/**
 	 * @return the builder with either lat and lon or city query parameter set to location.
 	 */
 	private UriBuilder addLocationQueryParam(UriBuilder builder, Location location) {
@@ -175,7 +98,7 @@ public class DwdWeatherServiceSourceAdapter implements SourceAdapter {
 		switch (time) {
 			case "forecast"://TODO
 			default:
-				return new CurrentResponseParser();
+				return new CurrentResponseParser(location);
 		}
 	}
